pip_requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
cycler==0.10.0
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
pandas==0.25.3
Pillow==6.2.1
protobuf==3.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.13.0
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
tensorflow-privacy==0.2.2
termcolor==1.1.0
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
Werkzeug==0.16.0
wrapt==1.11.2
